Update Events page a bit (#2180)
## Description

Howdy,

- I updated the UI of EventCard to match designs,
- made the events scroll horizontally, as per design,
- added a scroll-timeline + mask-image based fade to the scrollview,
similar to what we're using on landings, but fancier implementation as
we don't have as much negative space to render the fade gradient the
simpler way. This is maybe a 10th time where I used scroll-timeline,
it's pretty rad, my friends, good stuff.
- added GraphQLConf 2025 to past events
- and GraphQL Day at APIDays to upcoming events.

cc @Urigo you promised me a review today on the call :)

### Differences from designs

I made the spacing in the cards a little bit smaller because we don't
have as much room here in the docs as in the Events landing page they
were initially designed for.


https://github.com/user-attachments/assets/5ca41ec3-3d12-4a6c-a86e-d8e2fddd4cd5