bring 404 page closer to the designs (#2197)
## Description

Currently the 404 page doesn't respect light mode / dark mode (as it
doesn't use next-themes, app layout or nextra).

I just wanted to spend 5 minutes to improve it incrementally, and not
there yet, but we're closer.

To make it match designs, we'll need to add a mask-blurred stripes
decoration and render the footer and navbar (not super trivial as the
navbar takes props from Nextra.)