How to join a working group (#2158)
We had many people ask how to get involved in the working groups at
GraphQL Conference. Many times, I was told that they had tried to join
before but there were no easy steps to follow when it came to signing
the required CLA.

We even had questions asked about it during workshops, but because we
had all filled it out a long time ago it was easy to forget that it can
be a convoluted process.

---------

Co-authored-by: Benjie <benjie@jemjie.com>