new tools and libraries filters (#2177)
## Description

This PR adds a sidebar to the Tools and Libraries page, improves card
styles (even heights, no jagged edges), makes the card footer neutral
color (with a license icon for smaller screens instead of a word
"license") and adds


https://github.com/user-attachments/assets/57cb18bf-5894-4c4a-a28c-51c3a18f76c7

https://github.com/user-attachments/assets/d2c40fea-c2f2-4257-8513-5c2e06e53b6d

### Disclaimer

This PR has a trade-off. I needed to get rid of Headless UI from the
Navbar because the hydration error didn't let me code (like actually
blocked the dev server from doing anything productive), so we end up in
the state where we depend on a few components from Headless UI (also
because of Nextra) and few components from Base UI. I got rid of the
dependency on Radix RadioGroup so we end up with two headless component
libraries like before, but eventually it would be great to use just one.

To avoid making this PR larger I added Playwright tests to for the
Headless UI-related components and started migrating them in
https://github.com/graphql/graphql.github.io/pull/2182