Fix the navbar submenu in 404 page (#2222)
## Description

Currently, we render only the last segment of path in `href` in navbar
submenu links.

mea culpa, I tested if it opens, and I tested the other links, but I
didn't notice until I went back to work on `/community/events` page :/

This is currently only broken in 404 page, so it's not huge, but it can
be pretty frustrating to click a link and try to get out and land on 404
page again.