Use opacity transition instead of display:none (#2223)
## Background

I'm hiding the info message on keydown in the codemirror because it
sorta overlaps with the space for the query, and if you've seen and
started writing it you don't need it that much anymore.

Not my best work, but it's okay before the designers have an idea how to
compose it nicely.

## Description

This PR changes from `hidden` to `opacity-0` class to reduce the layout
shift and make it more pleasing to the eyes.