New events and meetups (#2230)
## Description

Refreshed the events page according to Figma.

I'll do a follow up PR with the `Add an event`, issue template, and
possibly "load more" instead of a scrollview, because this is big enough
already and I'd rather keep the convo about the event proposal issue
template separate from the code review of this page.


https://github.com/user-attachments/assets/c6c38a21-f30b-4afc-85b3-1f65f1a04991

---------

Co-authored-by: Jonathan Brennan <jonathanawesome@users.noreply.github.com>