blog: December Ambassador cohort update (#2272)
<!--
  Thanks for making a pull request!

  Before submitting, please read our contributing guidelines:

https://github.com/graphql/graphql.github.io/blob/source/CONTRIBUTING.md

  Have any questions?
Feel free to ask in this PR or you can also find us on the #website
channel on the GraphQL Slack (invite link available in CONTRIBUTING.md)
-->

## Description

<!-- Write a brief description of the changes introduced by this PR -->

An article for our new Ambassador cohort. 

I did some rejig of components to allow the Ambassador info card to
display just one cohort at a time and with or without social media
details.

---------

Co-authored-by: Benjie <benjie@jemjie.com>