Ambassador page: December cohort update (#2270)
<!--
  Thanks for making a pull request!

  Before submitting, please read our contributing guidelines:

https://github.com/graphql/graphql.github.io/blob/source/CONTRIBUTING.md

  Have any questions?
Feel free to ask in this PR or you can also find us on the #website
channel on the GraphQL Slack (invite link available in CONTRIBUTING.md)
-->


## Description

We have a new cohort of Ambassadors to welcome! Please generate a
preview which I can point our Ambassadors to so they can check how their
card looks and their social media links.

**Edit** Preview here:
https://graphql-github-io-git-fork-jemgi-8480cc-the-graph-ql-foundation.vercel.app/community/ambassadors/

Note that any corrections wont show up on the preview unless the preview
is generated again.